break up calculations for reports into seperate files
