+++ src/app/setup-wizard/data-collection/on-site-assessment/on-site-assessment-results/on-site-assessment-results.component.ts
@@ -7,8 +7,8 @@ import { NonEnergyBenefitsIdbService } from 'src/app/indexed-db/non-energy-benef
 import { IdbAssessment } from 'src/app/models/assessment';
 import { IdbEnergyOpportunity } from 'src/app/models/energyOpportunity';
 import { IdbNonEnergyBenefit } from 'src/app/models/nonEnergyBenefit';
-import { AssessmentReport, getAssessmentReport } from 'src/app/models/reports';
 import { KeyPerformanceMetric } from 'src/app/shared/constants/keyPerformanceMetrics';
+import { AssessmentReport, getAssessmentReport } from 'src/app/shared/reports/calculations/assessmentReport';
 
 @Component({
   selector: 'app-on-site-assessment-results',
@@ -41,7 +41,7 @@ export class OnSiteAssessmentResultsComponent {
     this.assessmentSub.unsubscribe();
   }
 
-  setAssessmentReport() {
+  setAssessmentReport(){
     let allEnergyOpportunities: Array<IdbEnergyOpportunity> = this.energyOpportunityIdbService.energyOpportunities.getValue();
     let allNonEnergyBenefits: Array<IdbNonEnergyBenefit> = this.nonEnergyBenefitIdbService.nonEnergyBenefits.getValue();
     let companyPerformanceMetrics: Array<KeyPerformanceMetric> = this.keyPerformanceIndicatorIdbService.getCompanyKeyPerformanceMetrics(this.assessment.companyId);
