+++ src/app/shared/reports/assessment-report/assessment-savings-table/assessment-savings-table.component.ts
@@ -1,7 +1,6 @@
 import { Component, Input } from '@angular/core';
 import { IconDefinition, faWeightHanging } from '@fortawesome/free-solid-svg-icons';
-import { IdbAssessment } from 'src/app/models/assessment';
-import { AssessmentReport } from 'src/app/models/reports';
+import { AssessmentReport } from '../../calculations/assessmentReport';
 
 @Component({
   selector: 'app-assessment-savings-table',
